sync/requirements.txt
google-cloud-storage
watchdog
